Feat #67 뉴스 검색 수정 및 인터페이스 생성
